requirements.txt
pytest==8.0.0
appium-python-client==3.1.1

